requirements.txt
fastapi
uvicorn
python-dotenv
motor
pydantic
google-genai
PyPDF2